requirements.txt
absl-py==2.3.1
alembic==1.17.2
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.12.0
libclang==18.1.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.3.5
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.44
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.1
